--- src/components/ReactGridLayout/GridItems.js
@@ -1,68 +1,0 @@
-<Card key={"1"} style={{overflow: 'none', backgroundColor:muiTheme.palette.primary2Color,
-                            }}>
-  <CardHeader
-    title="Asset Title"
-    subtitle="Slugline"
-    actAsExpander={true}
-    showExpandableButton={true}
-    />
-  <CardText expandable={true}>
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-    Donec mattis pretium massa. 
-  </CardText>
-  <CardActions expandable={true}>
-    <FlatButton label="Action1" />
-    <FlatButton label="Action2" />
-  </CardActions>
-</Card>
-<Card key={"2"} style={{overflow: 'none', backgroundColor:muiTheme.palette.primary2Color,
-                }}>
-  <CardHeader
-    title="Asset Title"
-    subtitle="Slugline"
-    actAsExpander={true}
-    showExpandableButton={true}
-    />
-  <CardText expandable={true}>
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-    Donec mattis pretium massa. 
-  </CardText>
-  <CardActions expandable={true}>
-    <FlatButton label="Action1" />
-    <FlatButton label="Action2" />
-  </CardActions>
-</Card>
-<Card key={"3"} style={{overflow: 'none', backgroundColor:muiTheme.palette.primary2Color,
-                }}>
-  <CardHeader
-    title="Asset Title"
-    subtitle="Slugline"
-    actAsExpander={true}
-    showExpandableButton={true}
-    />
-  <CardText expandable={true}>
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-    Donec mattis pretium massa. 
-  </CardText>
-  <CardActions expandable={true}>
-    <FlatButton label="Action1" />
-    <FlatButton label="Action2" />
-  </CardActions>
-</Card>
-<Card key={"4"} style={{overflow: 'none', backgroundColor:muiTheme.palette.primary2Color,
-                            }}>
-  <CardHeader
-    title="Asset Title"
-    subtitle="Slugline"
-    actAsExpander={true}
-    showExpandableButton={true}
-    />
-  <CardText expandable={true}>
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-    Donec mattis pretium massa. 
-  </CardText>
-  <CardActions expandable={true}>
-    <FlatButton label="Action1" />
-    <FlatButton label="Action2" />
-  </CardActions>
-</Card>
