--- src/components/ReactGridLayout/GridLayouts.js
@@ -1,10 +1,0 @@
-var layout = [
-  {i: '1', x: 6, y: 0, w: 4, h: 4.2,isResizable:false},
-  {i: '2', x: 0, y: 0, w: 3, h: 1},
-  {i: '3', x: 4, y: 0, w: 3, h: 3},
-  {i: '4', x: 4, y: 0, w: 3, h: 2},
-//  {i: '-KvYtBBgQVirwaZI7jm-', x: 4, y: 0, w: 3, h: 2}
-
-];
-
-var layouts = {md:layout}
