--- src/containers/ReactGridLayout/ReactGridLayout.js
@@ -1,148 +1,0 @@
-import React from "react";
-import { WidthProvider, Responsive } from "react-grid-layout";
-import _ from "lodash";
-import { connect } from 'react-redux';
-import { setSimpleValue } from '../../store/simpleValues/actions';
-import { setPersistentValue } from '../../store/persistentValues/actions';
-import {withRouter} from 'react-router-dom';
-import { withFirebase } from 'firekit';
-import {Card, CardActions, CardHeader, CardMedia, CardTitle, CardText} from 'material-ui/Card';
-import FloatingActionButton from 'material-ui/FloatingActionButton';
-import FontIcon from 'material-ui/FontIcon';
-import { onLayoutChange } from '../../store/grids/actions';
-import muiThemeable from 'material-ui/styles/muiThemeable';
-import { injectIntl } from 'react-intl';
-
-const ResponsiveReactGridLayout = WidthProvider(Responsive);
-
-/**
- * This layout demonstrates how to use a grid with a dynamic number of elements.
- */
-class ReactGridLayout extends React.PureComponent {
-  static defaultProps = {
-    className: "layout",
-    cols: { lg: 12, md: 10, sm: 6, xs: 4, xxs: 2 },
-    rowHeight: 90
-  };
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      // this is where the item data comes in. 
-      items: [0, 1, 2, 3, 4, 5] // this is where the items go. 
-      //this is where the layout comes in 
-
-      .map(function(i, key, list) {
-        return {
-          i: i.toString(),
-          x: i * 2,
-          y: 0,
-          w: 3,
-          h: 1,
-          add: i === (list.length - 1).toString()
-        };
-      }),
-      newCounter: 0
-    };
-
-    this.onAddItem = this.onAddItem.bind(this);
-    this.onBreakpointChange = this.onBreakpointChange.bind(this);
-  }
-
-  createElement(el) {
-    const removeStyle = {
-      position: "absolute",
-      right: "2px",
-      top: 0,
-      cursor: "pointer"
-    };
-    const widgetStyle = {
-      backgroundColor: "black",
-    }
-    const {muiTheme} =this.props;
-    const i = el.add ? "+" : el.i;
-    return (
-      <Card style={widgetStyle} key={i} data-grid={el}>
-        
-        {el.add ? (
-          <span
-            className="add text"
-            style={widgetStyle}
-            onClick={this.onAddItem}
-            title="You can add an item by clicking here, too."
-          >
-            Add +
-          </span>
-        ) : (
-          <span className="text">{i}</span>
-        )}
-        <span
-          className="remove"
-          style={removeStyle}
-          onClick={this.onRemoveItem.bind(this, i)}
-        >
-          x
-        </span>
-
-      </Card>
-    );
-  }
-
-  onAddItem() {
-    /*eslint no-console: 0*/
-    console.log("adding", "n" + this.state.newCounter);
-    this.setState({
-      // Add a new item. It must have a unique key!
-      items: this.state.items.concat({
-        i: "n" + this.state.newCounter,
-        x: (this.state.items.length * 2) % (this.state.cols || 12),
-        y: Infinity, // puts it at the bottom
-        w: 3,
-        h: 1
-      }),
-      // Increment the counter to ensure key is always unique.
-      newCounter: this.state.newCounter + 1
-    });
-  }
-
-  // We're using the cols coming back from this to calculate where to add new items.
-  onBreakpointChange(breakpoint, cols) {
-    this.setState({
-      breakpoint: breakpoint,
-      cols: cols
-    });
-  }
-
-  onLayoutChange(layout) {
-    this.props.onLayoutChange(layout);
-    this.setState({ layout: layout });
-  }
-
-  onRemoveItem(i) {
-    console.log("removing", i);
-    this.setState({ items: _.reject(this.state.items, { i: i }) });
-  }
-
-  render() {
-    return (
-      <div>
-        <div style={{position: 'fixed', left: 18, zIndex:3, bottom: 18, }}>
-          <FloatingActionButton style={{zIndex:3}} secondary={true} onClick={this.onAddItem}>
-            <FontIcon className="material-icons" >add</FontIcon>
-          </FloatingActionButton>
-        </div>
-        <ResponsiveReactGridLayout
-          onLayoutChange={this.onLayoutChange}
-          onBreakpointChange={this.onBreakpointChange}
-          {...this.props}
-        >
-          {_.map(this.state.items, el => this.createElement(el))}
-        </ResponsiveReactGridLayout>
-
-      </div>
-    );
-  }
-}
-
-export default (injectIntl(muiThemeable()(withRouter(withFirebase(ReactGridLayout)))));
