--- src/containers/ReactGridLayout/SavesReactGridLayout.js
@@ -1,123 +1,0 @@
-import React from "react";
-import RGL, { WidthProvider } from "react-grid-layout";
-import firebase from 'firebase';
-import {Card, CardActions, CardHeader, CardMedia, CardTitle, CardText} from 'material-ui/Card';
-import _ from 'lodash';
-
-const ReactGridLayout = WidthProvider(RGL);
-const originalLayout = getFromLS("layout") || [];
-const widgetStyle = {
-  backgroundColor: "grey",
-}
-
-
-/**
- * This layout demonstrates how to sync to localstorage.
- */
-export default class SavesReactGridLayout extends React.PureComponent {
-  static defaultProps = {
-    className: "layout",
-    cols: 12,
-    rowHeight: 50,
-    onLayoutChange: function() {}
-  };
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      layout: JSON.parse(JSON.stringify(originalLayout))
-    };
-
-    this.onLayoutChange = this.onLayoutChange.bind(this);
-    this.resetLayout = this.resetLayout.bind(this);
-  }
-
-  resetLayout() {
-    this.setState({
-      layout: []
-    });
-  }
-
-  onLayoutChange(layout) {
-    /*eslint no-console: 0*/
-    saveToLS("layout", layout);
-    this.setState({ layout });
-    this.props.onLayoutChange(layout); // updates status display
-  }
-  
-
-
-  render() {
-    const { list, match, firebaseApp, key, watchList, widgets,  } = this.props;
-    const uid=match.params.uid;
-
-    let ref=firebaseApp.database().ref('/widgets')
-    watchList(ref);
-
-    
-
-    return (
-      <div>
-        <button onClick={this.resetLayout}>Reset Layout</button>
-
-
-        <ReactGridLayout
-          {...this.props}
-          layout={this.state.layout}
-          onLayoutChange={this.onLayoutChange}
-        >
-        
-        <Card style={widgetStyle} key="1" data-grid={{ w: 3, h: 1, x: 0, y: 0 }}>
-          <span className="text">Asset</span>
-        </Card>
-
-        <Card style={widgetStyle} key="2" data-grid={{ w: 3, h: 1, x: 2, y: 0 }}>
-          <span className="text">Asset 2</span>
-        </Card>
-        <Card  style={widgetStyle} key="3" data-grid={{ w: 3, h: 4, x: 4, y: 0 }}>
-          <span className="text">Asset 3</span>
-        </Card>
-        <Card style={widgetStyle} key="4" data-grid={{ w: 4, h: 5, x: 6, y: 0 }}>
-          <span className="text">Widget 1</span>
-        </Card>
-        <Card style={widgetStyle} key="5" data-grid={{ w: 2, h: 4, x: 8, y: 0 }}>
-          <span className="text">Widget 2 </span>
-        </Card>
-         
-        </ReactGridLayout>
-      </div>
-    );
-  }
-}
-
-
-
-
-
-
-function getFromLS(key) {
-  let ls = {};
-  if (global.localStorage) {
-    try {
-      ls = JSON.parse(global.localStorage.getItem("rgl-7")) || {};
-    } catch (e) {
-      /*Ignore*/
-    }
-  }
-  return ls[key];
-}
-
-function saveToLS(key, value) {
-  if (global.localStorage) {
-    global.localStorage.setItem(
-      "rgl-7",
-      JSON.stringify({
-        [key]: value
-      })
-    );
-  }
-}
-
-//module.exports = LocalStorageLayout;
-
