+++ src/store/widgets/actions.js
@@ -1,0 +1,9 @@
+import * as types from './types';
+
+export function onWidgetChange(widget){
+  return {
+    type: types.ON_WIDGET_CHANGED,
+    widget,
+    
+  };
+}
