+++ src/store/widgets/types.js
@@ -1,0 +1,3 @@
+const namespace='WIDGETS'
+
+export const ON_WIDGET_CHANGED = `${namespace}@ON_WIDGET_CHANGED`;
